Rename gk input readers
