+++ pyrokinetics/gk_code/__init__.py
@@ -1,8 +1,8 @@
-from .GKInput import GKInput, gk_inputs
-from .GKInputGS2 import GKInputGS2
-from .GKInputCGYRO import GKInputCGYRO
-from .GKInputGENE import GKInputGENE
-from .GKInputTGLF import GKInputTGLF
+from .gk_input import GKInput, gk_inputs
+from .gs2 import GKInputGS2
+from .cgyro import GKInputCGYRO
+from .gene import GKInputGENE
+from .tglf import GKInputTGLF
 
 gk_inputs["GS2"] = GKInputGS2
 gk_inputs["CGYRO"] = GKInputCGYRO
