+++ pyrokinetics/gk_code/gs2.py
@@ -13,7 +13,7 @@ from ..local_geometry import (
 from ..numerics import Numerics
 from ..templates import gk_templates
 from ..normalisation import ureg, SimulationNormalisation as Normalisation, convert_dict
-from .GKInput import GKInput
+from .gk_input import GKInput
 
 
 class GKInputGS2(GKInput):
